Refactoring of Company_advertisement.js to CompanyAdvertisement.js
